Add fast path for upvalues in `OP_GETTABUP`
The rationale is that `luaV_gettable` **seems** to be doing work that isn't strictly necessary in the case of upvalues. Additionally, we can ease inlining and probably help avoid register allocation carnage by simplifying this code path significantly.

There seems to be significant gains for e.g. the waves demo.

I believe this should only break the VM in terrible uses of `_ENV`, which I am tempted not to care about.